lock board
If we try to flip a second set of cards before the first set has
finished unflipping, it crushes our logic. We have to lock the board and
wait till cards finished unflipping. So, in case it is not a match we
lock the board. And only unlock it once the cards have been flipped.